Resolved stash conflicts

+++ Application/Services/AddressService.cs
@@ -1,4 +1,5 @@
 ﻿using Core.Entities;
+using Core.Exceptions;
 using Core.Interfaces;
 using Core.Interfaces.Repositories;
 using Core.Interfaces.Services;
@@ -18,16 +19,14 @@ namespace Application.Services
 
         public async Task<Address> GetAddressByUserIdAsync(int id)
         {
-            _loggerManager.LogInfo("Inside GetAddressByUserIdAsync");
             var address = await _addressRepository.GetAddressByUserIdAsync(id);
             if (address != null)
             {
-                _loggerManager.LogInfo("Return from GetAddressByUserIdAsync");
                 return address;
             }
 
             _loggerManager.LogError($"User with ID - {id} doesn't have an address!");
-            throw new NullReferenceException();
+            throw new NotFoundException($"User with ID - {id} doesn't have an address!");
         }
 
         public async Task<IEnumerable<Address>> GetAllAddressesAsync()
@@ -37,17 +36,41 @@ namespace Application.Services
 
         public async Task CreateAddressAsync(Address address)
         {
-            await _addressRepository.CreateAddressAsync(address);
+            var possibleAddressInTable = await _addressRepository.GetAddressByUserIdAsync(address.UserId);
+            if (possibleAddressInTable == null)
+            {
+                await _addressRepository.CreateAddressAsync(address);
+                await _addressRepository.SaveChangesAsync();
+            }
+
+            _loggerManager.LogError($"User with ID - {address.UserId} has already an address!");
+            throw new BadRequestException($"User with ID - {address.UserId} has already an address!");
         }
 
         public async Task UpdateAddressAsync(Address address)
         {
-            await _addressRepository.UpdateAddressAsync(address);
+            var portfolioInTable = await _addressRepository.GetAddressByUserIdAsync(address.UserId);
+            if (portfolioInTable != null)
+            {
+                await _addressRepository.UpdateAddressAsync(address);
+                await _addressRepository.SaveChangesAsync();
+            }
+
+            _loggerManager.LogError($"User with ID - {address.UserId} doesn't have an address!");
+            throw new NotFoundException($"User with ID - {address.UserId} doesn't have an address!");
         }
 
         public async Task DeleteAddressByUserIdAsync(int id)
         {
-            await _addressRepository.DeleteAddressByUserIdAsync(id);
+            var portfolioInTable = await _addressRepository.GetAddressByUserIdAsync(id);
+            if (portfolioInTable != null)
+            {
+                await _addressRepository.DeleteAddressByUserIdAsync(id);
+                await _addressRepository.SaveChangesAsync();
+            }
+
+            _loggerManager.LogError($"User with ID - {id} doesn't have an address!");
+            throw new NotFoundException($"User with ID - {id} doesn't have an address!");
         }
     }
 }
